Restructure files to have better management
