--- packages/picasso-lexical-rte/src/LexicalRTE/index.ts
@@ -1,4 +1,0 @@
-export { default } from './LexicalRTE'
-export { default as LinkPlugin } from './plugins/LinkPlugin';
-export { Heading } from './plugins/ToolbarPlugin';
-export { LinkNode } from './nodes/EditorNodes';
