+++ packages/picasso-lexical-rte/src/components/LexicalRTE/styles.ts
