+++ packages/picasso-lexical-rte/src/index.ts
@@ -1,6 +1,7 @@
 export {
   default as LexicalRTE,
-  LinkPlugin,
-  LinkNode,
-  Heading,
-} from './LexicalRTE'
+} from './components/LexicalRTE'
+
+export { default as LinkPlugin } from './plugins/LinkPlugin';
+export { Heading } from './plugins/ToolbarPlugin';
+export { LinkNode } from './nodes/EditorNodes';
