+++ packages/picasso-lexical-rte/src/plugins/FocusPlugin.ts
