+++ packages/picasso-lexical-rte/src/plugins/styles.ts
