+++ packages/picasso-lexical-rte/src/utils/getSelectedNode.ts
