+++ packages/picasso-lexical-rte/src/utils/index.ts
