+++ packages/picasso-lexical-rte/src/utils/validateUrl.ts
